Hackily fix interpolation returning bad results
The temporary (distance * (to - from)) is stored as a uint32_t, which
is too small to contain the result if the input type is uint32_t as
well.  However, was fine for smaller types, so this is only cropping up
now.  Solution for now is to just make the temporary result bigger.  See
also http://stackoverflow.com/questions/35051394 for, hopefully, a more
permanent solution.